requirements.txt
setuptools==62.0.0
setuptools-scm
wheel

xlsxwriter
openpyxl
pandas
Pillow